emit_expr: Add tests and move to emit module.
